pyproject.toml
[tool.poetry]
name = "book-agent"
version = "0.1.0"
description = "AI-Enhanced Interactive Book Agent"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.1"
uvicorn = {version = "^0.24.0", extras = ["standard"]}
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
langchain = "^0.0.332"
chromadb = "^0.4.22"
google-generativeai = "^0.3.1"
pydantic = "^2.5.0"
python-jose = "^3.3.0"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
python-multipart = "^0.0.6"
tiktoken = "^0.5.2"
pdfplumber = "^0.9.0"
python-docx = "^1.1.0"
ebooklib = "^0.18"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dev = "run_dev:main"
test = "pytest"
test-cov = "pytest --cov=backend.src --cov-report=html --cov-report=term"

[tool.pytest.ini_options]
testpaths = ["backend/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
]

[tool.coverage.run]
source = ["backend/src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/.env/*",
    "backend/src/main.py",  # Main entry point doesn't need coverage
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:"
]